download a piece

+++ src/util.hpp
@@ -25,5 +25,5 @@ int write_to_file(char *filename, std::vector<struct Piece> &piece);
 std::map<std::string, std::string> parse_magnet(const std::string& magnet);
 int handle_magnet_peers(const std::string announce_url, const std::string hash, std::vector<std::string> &ips, std::vector<uint16_t> &ports);
 int handle_magnet_handshake(const std::string ip, const uint16_t port, const std::string hash, int &metadata_id);
-json handle_magnet_info(const int sockfd, uint8_t metadata_id, unsigned int piece);
+json handle_magnet_info(const int sockfd, int metadata_id, unsigned int piece);
 #endif
